fix double float error

+++ frameworks/core/interfaces/native/implementation/accessibility_hover_event_accessor.cpp
@@ -71,7 +71,7 @@ void SetXImpl(Ark_AccessibilityHoverEvent peer, Ark_Float64 x)
     CHECK_NULL_VOID(info);
     auto localLocation = info->GetLocalLocation();
     const auto animation = localLocation.GetXAnimationOption();
-    auto value = Converter::Convert<float>(x);
+    auto value = Converter::Convert<double>(x);
     auto xConvert = PipelineBase::Vp2PxWithCurrentDensity(value);
     localLocation.SetX(xConvert, animation);
     info->SetLocalLocation(localLocation);
@@ -92,7 +92,7 @@ void SetYImpl(Ark_AccessibilityHoverEvent peer, Ark_Float64 y)
     CHECK_NULL_VOID(info);
     auto localLocation = info->GetLocalLocation();
     const auto animation = localLocation.GetYAnimationOption();
-    auto value = Converter::Convert<float>(y);
+    auto value = Converter::Convert<double>(y);
     auto yConvert = PipelineBase::Vp2PxWithCurrentDensity(value);
     localLocation.SetY(yConvert, animation);
     info->SetLocalLocation(localLocation);
@@ -114,7 +114,7 @@ void SetDisplayXImpl(Ark_AccessibilityHoverEvent peer, Ark_Float64 displayX)
     CHECK_NULL_VOID(info);
     auto screenLocation = info->GetScreenLocation();
     const auto animation = screenLocation.GetXAnimationOption();
-    auto value = Converter::Convert<float>(displayX);
+    auto value = Converter::Convert<double>(displayX);
     auto xConvert = PipelineBase::Vp2PxWithCurrentDensity(value);
     screenLocation.SetX(xConvert, animation);
     info->SetScreenLocation(screenLocation);
@@ -136,7 +136,7 @@ void SetDisplayYImpl(Ark_AccessibilityHoverEvent peer, Ark_Float64 displayY)
     CHECK_NULL_VOID(info);
     auto screenLocation = info->GetScreenLocation();
     const auto animation = screenLocation.GetYAnimationOption();
-    auto value = Converter::Convert<float>(displayY);
+    auto value = Converter::Convert<double>(displayY);
     auto yConvert = PipelineBase::Vp2PxWithCurrentDensity(value);
     screenLocation.SetY(yConvert, animation);
     info->SetScreenLocation(screenLocation);
@@ -158,7 +158,7 @@ void SetWindowXImpl(Ark_AccessibilityHoverEvent peer, Ark_Float64 windowX)
     CHECK_NULL_VOID(info);
     auto globalLocation = info->GetGlobalLocation();
     const auto animation = globalLocation.GetXAnimationOption();
-    auto value = Converter::Convert<float>(windowX);
+    auto value = Converter::Convert<double>(windowX);
     auto xConvert = PipelineBase::Vp2PxWithCurrentDensity(value);
     globalLocation.SetX(xConvert, animation);
     info->SetGlobalLocation(globalLocation);
@@ -180,7 +180,7 @@ void SetWindowYImpl(Ark_AccessibilityHoverEvent peer, Ark_Float64 windowY)
     CHECK_NULL_VOID(info);
     auto globalLocation = info->GetGlobalLocation();
     const auto animation = globalLocation.GetYAnimationOption();
-    auto value = Converter::Convert<float>(windowY);
+    auto value = Converter::Convert<double>(windowY);
     auto yConvert = PipelineBase::Vp2PxWithCurrentDensity(value);
     globalLocation.SetY(yConvert, animation);
     info->SetGlobalLocation(globalLocation);
